2 sperate files have been created for each row inside the main projects page
